refactor: rename TopBar to Header
