+++ src/components/layout/index.tsx
@@ -3,7 +3,7 @@ import { graphql, useStaticQuery } from 'gatsby';
 import Footer from '../footer';
 import Menu from '../../../lib/SSW.MegaMenu/menu/menu';
 import MobileMenu from '../../../lib/SSW.MegaMenu/mobile-menu/mobile-menu';
-import TopBar from '../topBar';
+import Header from '../header';
 import '@fortawesome/fontawesome-svg-core/styles.css';
 
 type LayoutProps = {
@@ -46,7 +46,7 @@ const Layout = ({ pageTitle, children }: LayoutProps) => {
                 <div style={{ backgroundColor: 'white' }}>
                     <div className="flex flex-col min-h-screen main-container">
                         <title>{pageTitle}</title>
-                        <TopBar />
+                        <Header />
                         <Menu
                             onClickToggle={() => actionOnToggleClick()}
                         ></Menu>
